2nd Progress Check
HTML/CSS: Started on changing the header and body to make it fit inside the desktop window, then had to fix a bunch of text alignments on the footer and everywhere else. Started on the top news html site.

JS: Started with a variable, may continue to add more to the function of the website.

JAVA: May be added to store information and create a algorithm that caters to viewer's preferences based on view history.